Adding some corrections
