+++ app/models/lesson.rb
@@ -10,6 +10,6 @@
 #
 
 class Lesson < ApplicationRecord
-  validates :title, presence: true, length: { maximum: 50 }
+  validates :title, presence: true, allow_blank: false, length: { maximum: 50 }
   validates :description, length: { maximum: 300 }
 end
