+++ spec/support/json_helper.rb
@@ -1,0 +1,9 @@
+module JsonHelper
+  def json_response
+    if (json = JSON.parse(response.body)).is_a?(Array)
+      json.map!(&:with_indifferent_access)
+    else
+      json.with_indifferent_access
+    end
+  end
+end
